back-end/requirements.txt
asgiref==3.2.3
cffi==1.13.2
cryptography==2.8
Django==3.0.3
django-cors-headers==3.2.1
django-environ==0.4.5
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djoser==2.0.3
joblib==0.14.1
mysqlclient==1.4.6
numpy==1.18.1
pandas==0.25.3
py4j==0.10.7
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
pyspark==2.4.5
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
scikit-learn==0.22.1
scikit-surprise==1.1.0
scipy==1.4.1
six==1.13.0
sklearn==0.0
sqlparse==0.3.0
surprise==0.1
